sharesansar/requirements.txt
selenium==4.28.1
beautifulsoup4==4.12.3
rich==13.9.4
python-dotenv==1.0.1
pandas==2.2.3
sqlalchemy==2.0.37